Implements folder structure for components
