+++ src/App.js
@@ -1,22 +1,22 @@
 import React, { Component, createRef } from 'react'
 import './semantic/dist/semantic.min.css'
 import Navbar from './Components/Navbar'
-import Menu from './Components/Menu'
+import Menu from './Components/Menu/Menu'
 import Search from './Components/Search'
 import SearchResults from './Components/SearchResults'
-import AboutUs from './Components/AboutUs'
-import Blog from './Components/Blog'
-import ContactUs from './Components/ContactUs'
-import Faq from './Components/Faq'
-import Legal from './Components/Legal'
-import Login from './Components/Login'
-import SignUp from './Components/SignUp'
-import SignupSuccess from './Components/SignupSuccess'
-import PasswordReset from './Components/PasswordReset'
-import ChangePassword from './Components/ChangePassword'
-import PasswordResetSuccess from './Components/PasswordResetSuccess'
+import AboutUs from './Components/Menu/AboutUs'
+import Blog from './Components/Menu/Blog'
+import ContactUs from './Components/Menu/ContactUs'
+import Faq from './Components/Menu/Faq'
+import Legal from './Components/Menu/Legal'
+import Login from './Components/Authentication/Login'
+import SignUp from './Components/Authentication/SignUp'
+import SignupSuccess from './Components/Authentication/SignupSuccess'
+import PasswordReset from './Components/Authentication/PasswordReset'
+import ChangePassword from './Components/Authentication/ChangePassword'
+import PasswordResetSuccess from './Components/Authentication/PasswordResetSuccess'
 import UserPage from './Components/UserPage'
-import ScrollToTop from './Components/ScrollToTop'
+import ScrollToTop from './Modules/ScrollToTop'
 import { Container, Sticky, Sidebar } from 'semantic-ui-react'
 import { connect } from 'react-redux'
 import { Switch, Route } from 'react-router-dom'
