+++ src/Components/HostProfileView/Reviews.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Segment, Header } from 'semantic-ui-react'
+
+const Reviews = () => {
+  return (
+    <>
+      <Header as='h3' style={{ 'textAlign': 'left' }}>
+        Reviews
+      </Header>
+      <Segment placeholder style={{ 'textAlign': 'center' }}>
+        Here will be rewiews
+      </Segment>
+    </>
+  )
+}
+
+export default Reviews
