+++ src/Components/Menu/Blog.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 import { Header } from 'semantic-ui-react'
-import WIP from './WorkInProgress'
+import WIP from '../ReusableComponents/WorkInProgress'
 
-class Faq extends Component {
+class Blog extends Component {
   render() {
     return (
       <div className='content-wrapper' >
         <Header as='h1'>
-          FAQ
+          Blog
         </Header>
 
         <WIP />
@@ -17,4 +17,4 @@ class Faq extends Component {
   }
 }
 
-export default Faq
+export default Blog
