+++ src/Components/Menu/Faq.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 import { Header } from 'semantic-ui-react'
-import WIP from './WorkInProgress'
+import WIP from '../ReusableComponents/WorkInProgress'
 
-class Blog extends Component {
+class Faq extends Component {
   render() {
     return (
       <div className='content-wrapper' >
         <Header as='h1'>
-          Blog
+          FAQ
         </Header>
 
         <WIP />
@@ -17,4 +17,4 @@ class Blog extends Component {
   }
 }
 
-export default Blog
+export default Faq
