--- src/Components/Reviews.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import { Segment, Header } from 'semantic-ui-react'
-
-const Reviews = () => {
-  return (
-    <>
-    <Header as='h3' style={{'textAlign': 'left'}}>
-      Reviews
-    </Header>
-    <Segment placeholder style={{'textAlign': 'center'}}>
-      Here will be rewiews
-    </Segment>
-    </>
-  )
-}
-
-export default Reviews
