+++ src/Components/SearchResults.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux'
 import Geocode from 'react-geocode'
 import { bookingSearch, getBookingLength } from '../Modules/booking'
 import List from './List'
-import GoogleMap from './GoogleMap'
-import HostProfileView from './HostProfileView'
+import GoogleMap from './Map/GoogleMap'
+import HostProfileView from './HostProfileView/HostProfileView'
 import moment from 'moment'
 import axios from 'axios'
 import Popup from 'reactjs-popup'
