+++ src/index.js
@@ -12,7 +12,7 @@ import axios from 'axios'
 const development = 'http://localhost:3007'
 const production = process.env.REACT_APP_API_ENDPOINT
 
-axios.defaults.baseURL = production
+axios.defaults.baseURL = development
 
 const store = configureStore()
 verifyCredentials(store)
