+++ src/reduxTokenAuthConfig.js
@@ -5,7 +5,7 @@ const production = `${process.env.REACT_APP_API_ENDPOINT}/api/v1/auth`
 const development = 'http://localhost:3007/api/v1/auth'
 
 const config = {
-  authUrl: production,
+  authUrl: development,
   userAttributes: {
     id: 'id',
     uid: 'uid',
